Added employee dashboard

+++ Backend/controllers/userController.js
@@ -61,7 +61,12 @@ const signUp = async (req, res) => {
 }
 
 const getAllEmployee = (req, res) => {
-    const sql = "SELECT * FROM employee";
+    const sql = `
+        SELECT e.id, e.username, e.email, r.id as role_id, r.name as role
+        FROM employee AS e
+        JOIN role AS r ON e.role_id = r.id
+        ORDER BY e.id ASC
+    `
     db.query(sql, (err, results) => {
         if (err) {
             return res.status(500).json({
@@ -102,25 +107,94 @@ const getEmployeeById = (req, res) => {
 
 const deleteEmployee = (req, res) => {
     const { id } = req.params;
-    const sql = `DELETE FROM employee WHERE id= ${id}`;
-    db.query(sql, [id], (err, results) => {
+
+    const checkEmployeeQuery = "SELECT * FROM employee WHERE id = ?";
+    db.query(checkEmployeeQuery, [id], (err, results) => {
         if (err) {
             return res.status(500).json({
                 status: false,
                 error: err.message
             });
         }
-        if (results.affectedRows === 0) {
+        if (results.length === 0) {
             return res.status(404).json({
                 status: false,
                 error: "Employee not found"
             });
         }
-        res.status(200).json({
-            status: true,
-            message: "Employee deleted successfully"
+        if (results[0].role_id === 1) {
+            return res.status(403).json({
+                status: false,
+                error: "Cannot delete admin employee"
+            });
+        }
+
+        const sql = `DELETE FROM employee WHERE id= ${id}`;
+        db.query(sql, [id], (err, results) => {
+            if (err) {
+                return res.status(500).json({
+                    status: false,
+                    error: err.message
+                });
+            }
+            if (results.affectedRows === 0) {
+                return res.status(404).json({
+                    status: false,
+                    error: "Employee not found"
+                });
+            }
+            res.status(200).json({
+                status: true,
+                message: "Employee deleted successfully"
+            });
         });
     });
+}
+
+const updateEmployeeRole = (req, res) => {
+    const { id } = req.params;
+    const { role_id } = req.body;
+
+    const checkQuery = "SELECT * FROM employee WHERE id = ?";
+    db.query(checkQuery, [id], (err, results) => {
+        if (err) {
+            return res.status(500).json({
+                status: false,
+                error: err.message
+            });
+        }
+        if (results.length === 0) {
+            return res.status(404).json({
+                status: false,
+                error: "Employee not found"
+            });
+        }
+        if (results[0].role_id === 1) {
+            return res.status(403).json({
+                status: false,
+                error: "Cannot update admin employee role"
+            });
+        }
+        const sql = `UPDATE employee SET role_id = ? WHERE id = ?`;
+        db.query(sql, [role_id, id], (err, results) => {
+            if (err) {
+                return res.status(500).json({
+                    status: false,
+                    error: err.message
+                });
+            }
+            if (results.affectedRows === 0) {
+                return res.status(404).json({
+                    status: false,
+                    error: "Employee not found"
+                });
+            }
+            res.status(200).json({
+                status: true,
+                message: "Employee role updated successfully"
+            })
+        })
+    })
 }
 
 module.exports = {
@@ -128,5 +202,6 @@ module.exports = {
     signUp,
     getAllEmployee,
     getEmployeeById,
-    deleteEmployee
+    deleteEmployee,
+    updateEmployeeRole
 }
